+++ packages/featureserver/src/geometry-map.js
@@ -1,0 +1,8 @@
+module.exports = {
+  Point: 'esriGeometryPoint',
+  MultiPoint: 'esriGeometryMultipoint',
+  LineString: 'esriGeometryPolyline',
+  MultiLineString: 'esriGeometryPolyline',
+  Polygon: 'esriGeometryPolygon',
+  MultiPolygon: 'esriGeometryPolygon'
+}