+++ packages/featureserver/src/utils.js
@@ -1,4 +1,5 @@
 const esriExtent = require('esri-extent')
+const geometryMap = require('./geometry-map')
 
 /**
  * if we have no extent, but we do have features, then it should be Table
@@ -8,7 +9,8 @@ const esriExtent = require('esri-extent')
  * @return {boolean}
  */
 function isTable (json, data) {
-  var noExtent = (!json.fullExtent || !json.fullExtent.xmin || !json.fullExtent.ymin) || json.fullExtent.xmin === Infinity
+  if (json.geometryType || (json.metadata && json.metadata.geometryType)) return false
+  var noExtent = !json.fullExtent || !json.fullExtent.xmin || !json.fullExtent.ymin || json.fullExtent.xmin === Infinity
   var hasFeatures = data.features || data[0].features
   if (noExtent && !hasFeatures) return true
   else return false
@@ -25,9 +27,10 @@ const esriGeomTypes = {
   point: 'esriGeometryPoint'
 }
 
-function getGeomType (geojson) {
+function getGeomType (geojson = {}) {
   // TODO this should find the first non-null geometry
-  if (!geojson || !geojson.features || !geojson.features[0]) return undefined
+  if (geojson.metadata && geojson.metadata.geometryType) return geometryMap[geojson.metadata.geometryType]
+  if (!geojson.features || !geojson.features[0]) return undefined
   const feature = geojson.features[0]
   if (!feature || !feature.geometry || !feature.geometry.type) return undefined
   const type = esriGeomTypes[feature.geometry.type.toLowerCase()]