+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Support for passing a server extent e.g. [[-180, -90], [180, 90]]
+* Support passed geometryType passed in layer metadata
+
 ## [2.0.0] - 2017-06-01
 ### Added
 * /FeatureServer supports multiple layers